pyproject.toml
[tool.poetry]
name = "election-edu-bot"
version = "0.1.0"
description = "Election Education Bot"
authors = ["Antigravity <antigravity@google.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
google-cloud-aiplatform = "^1.35.0"
google-generativeai = "^0.8.0"
google-genai = "^1.0.0"
google-cloud-firestore = "^2.13.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
